+++ algoliasearch-common/src/main/java/com/algolia/search/Index.java
@@ -1796,11 +1796,11 @@ public class Index<T>
     private String name;
     private String createdAt;
     private String updatedAt;
-    private Integer entries;
-    private Integer dataSize;
-    private Integer fileSize;
-    private Integer lastBuildTimeS;
-    private Integer numberOfPendingTask;
+    private Long entries;
+    private Long dataSize;
+    private Long fileSize;
+    private Long lastBuildTimeS;
+    private Long numberOfPendingTask;
     private Boolean pendingTask;
 
     public String getName() {
@@ -1815,23 +1815,23 @@ public class Index<T>
       return updatedAt;
     }
 
-    public Integer getEntries() {
+    public Long getEntries() {
       return entries;
     }
 
-    public Integer getDataSize() {
+    public Long getDataSize() {
       return dataSize;
     }
 
-    public Integer getFileSize() {
+    public Long getFileSize() {
       return fileSize;
     }
 
-    public Integer getLastBuildTimeS() {
+    public Long getLastBuildTimeS() {
       return lastBuildTimeS;
     }
 
-    public Integer getNumberOfPendingTask() {
+    public Long getNumberOfPendingTask() {
       return numberOfPendingTask;
     }
